AuthSlice created
