+++ src/pages/login/SignIn.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import * as Yup from "yup";
 import { FaUserAlt, FaLock } from "react-icons/fa";
 import { useDispatch, useSelector } from "react-redux";
-import { UserLoginAction } from "../../redux/slices/UserSlice";
 import { useFormik } from "formik";
 import { Link } from "react-router-dom";
+import { UserLoginAction } from "../../redux/slices/AuthSlice";
 
 //Form schema
 const formSchema = Yup.object({
