+++ src/redux/store/Store.js
@@ -1,4 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
+import AuthSlice from "../slices/AuthSlice";
 import PostSlice from "../slices/PostSlice";
 import UserSlice from "../slices/UserSlice";
 
@@ -6,6 +7,7 @@ const store = configureStore({
   reducer: {
     user: UserSlice,
     post: PostSlice,
+    auth: AuthSlice,
   },
 });
 // The store now has redux-thunk added and the Redux DevTools Extension is turned on
